🦹‍♀️ refactor(Typescript): 命名空间选择接口修改
